llm-sm-selection/requirements.txt
pyyaml
openai
anthropic
google-generativeai
ipykernel
pandas
scikit-learn
matplotlib
seaborn
